Update AudioMonitor.vhd
We don't have to explicitly put parsed_data as the output for every state, since if parsed_data isn't change, the value is kept. The only time it would change is if a clap is detected or if CS is active. When CS is active, the parsed_data is set to 0.